@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.8.6+30
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: "3.6.1"
 
 dependencies:
   flutter:
@@ -41,15 +41,15 @@ dependencies:
   multi_select_flutter: ^4.0.0
   # flutter_datetime_picker: 1.5.1
   modal_bottom_sheet: ^3.0.0-pre
-  image_cropper: 1.5.0
+  image_cropper: 9.0.0
   image_picker: ^0.8.6
-  file_picker: 5.2.5
-  crop_image: 1.0.2
+  file_picker: 9.0.2
+  crop_image: 1.0.16
   image: ^3.1.3
   path: ^1.8.0
-  uri_to_file: ^0.2.0
+  uri_to_file: 1.0.0
   uni_links: ^0.5.1
-  filesystem_picker: ^3.0.0-beta.1
+  filesystem_picker: 4.1.0
 
 
 dev_dependencies:
