@@ -190,7 +190,7 @@ Future<String?> displayTextInputDialog(BuildContext context,
                               fontSize: 12,
                               color: Theme.of(context)
                                   .textTheme
-                                  .bodyText1
+                                  .bodyLarge
                                   ?.color
                                   ?.withOpacity(.5)),
                         ),
