@@ -239,7 +239,7 @@ class ViewInfoCard extends StatelessWidget {
                           fontSize: 13,
                           color: Theme.of(context)
                               .textTheme
-                              .bodyText1!
+                              .bodyLarge!
                               .color!
                               .withAlpha(0xCC),
                         ),
