@@ -112,7 +112,7 @@ class _DownloadComicsScreenState extends State<DownloadComicsScreen> {
         return Container(
           width: constraints.maxWidth,
           padding: const EdgeInsets.all(15),
-          color: (Theme.of(context).textTheme.bodyText1?.color ?? Colors.black)
+          color: (Theme.of(context).textTheme.bodyLarge?.color ?? Colors.black)
               .withOpacity(.05),
           child: Text(
             text,
