@@ -24,7 +24,7 @@ class DownloadInfoCard extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     var theme = Theme.of(context);
-    var textColor = theme.textTheme.bodyText1!.color!;
+    var textColor = theme.textTheme.bodyLarge!.color!;
     var textColorAlpha = textColor.withAlpha(0x33);
     var textColorSummary = textColor.withAlpha(0xCC);
     var titleStyle = TextStyle(
